Implement box discovery
- Update Kotlin to 1.1.4-2
- Implements basic box discovery with NSDManager
- Format files
- Format headers
- Abstract Volley to own function
