+++ app/build.gradle
@@ -1,4 +1,4 @@
-/**
+/*
  * This file is part of FreedomBox.
  *
  * FreedomBox is free software: you can redistribute it and/or modify
