+++ app/src/main/AndroidManifest.xml
@@ -8,14 +8,14 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
-        android:name=".Applications.FreedomBoxApp"
+        android:name=".applications.FreedomBoxApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/FreedomBoxTheme">
 
-        <activity android:name=".Views.Activities.MainActivity">
+        <activity android:name=".views.activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -23,10 +23,10 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".Views.Activities.LauncherActivity">
+        <activity android:name=".views.activities.LauncherActivity">
         </activity>
 
-        <activity android:name=".Views.Activities.DiscoveryActivity" />
+        <activity android:name=".views.activities.DiscoveryActivity" />
 
     </application>
 </manifest>
