--- app/src/main/java/org/freedombox/freedombox/SVGModules/SvgDecoder.kt
@@ -1,35 +1,0 @@
-/**
- *  This file is part of FreedomBox.
- *
- * FreedomBox is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * FreedomBox is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with FreedomBox. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.freedombox.freedombox.SVGModules
-
-import com.bumptech.glide.load.ResourceDecoder
-import com.bumptech.glide.load.resource.SimpleResource
-import com.caverock.androidsvg.SVG
-import com.caverock.androidsvg.SVGParseException
-import java.io.InputStream
-
-class SvgDecoder : ResourceDecoder<InputStream, SVG> {
-
-    override fun decode(source: InputStream?, width: Int, height: Int) = try {
-        SimpleResource<SVG>(SVG.getFromInputStream(source))
-    } catch (e: SVGParseException) {
-        null
-    }
-
-    override fun getId() = "org.freedombox.freedombox.svg"
-}
