--- app/src/main/java/org/freedombox/freedombox/SVGModules/SvgDrawableTranscoder.kt
@@ -1,37 +1,0 @@
-/**
- *  This file is part of FreedomBox.
- *
- * FreedomBox is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * FreedomBox is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with FreedomBox. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.freedombox.freedombox.SVGModules
-
-import android.graphics.drawable.PictureDrawable
-import com.bumptech.glide.load.engine.Resource
-import com.bumptech.glide.load.resource.SimpleResource
-import com.bumptech.glide.load.resource.transcode.ResourceTranscoder
-import com.caverock.androidsvg.SVG
-
-class SvgDrawableTranscoder: ResourceTranscoder<SVG, PictureDrawable> {
-
-    override fun getId() = "org.freedombox.freedombox"
-
-    override fun transcode(toTranscode: Resource<SVG>): Resource<PictureDrawable> {
-        val svg = toTranscode.get()
-        val picture = svg.renderToPicture()
-        val drawable = PictureDrawable(picture)
-
-        return SimpleResource<PictureDrawable>(drawable)
-    }
-}
