+++ app/src/main/java/org/freedombox/freedombox/views/activities/BaseActivity.kt
@@ -1,0 +1,43 @@
+/*
+ * This file is part of FreedomBox.
+ *
+ * FreedomBox is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * FreedomBox is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with FreedomBox. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.freedombox.freedombox.views.activities
+
+import android.os.Bundle
+import android.support.v4.app.Fragment
+import android.support.v7.app.AppCompatActivity
+import org.freedombox.freedombox.R
+
+abstract class BaseActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContentView(R.layout.base)
+    }
+
+    fun loadFragment(fragmentContainerId: Int, fragment: Fragment, addToBackStack: Boolean = false) {
+        val transaction = this.supportFragmentManager.beginTransaction()
+
+        transaction.replace(fragmentContainerId, fragment, fragment.javaClass.name)
+
+        if (addToBackStack) {
+            transaction.addToBackStack(fragment.javaClass.name)
+        }
+
+        transaction.commit()
+    }
+}
